packages/core/pyproject.toml
[project]
name = "aiai-core"
version = "0.1.0"
description = "Shared types, config loading, and structured logging for aiai"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aiai_core"]
